Added maxBodySize option which required me to make factories for middleware.
I updated tests, added new tests for new test cases, updated the docs,
and added bytes for parsing the maxBodySize if it is a string such as 100kb. An
addition unrelated change was I added a discardBody middleware and used
it to ensure the body gets consumed if an error occurs or it simply
wasn't read.

The end goal of all of this was so we could enforce a max body size.